backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai>=1.6.1,<2.0.0 
langchain-openai==0.0.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4; python_version < '3.12'
PyMuPDF==1.23.6
python-docx==0.8.11
pdfplumber==0.9.0
unstructured==0.10.30
numpy>=1.26.0
python-dotenv==1.0.0